refactor: change atom components to specific folder
