+++ components/PageNavigator.tsx
@@ -1,4 +1,4 @@
-import PrimaryButton from "./PrimaryButton"
+import PrimaryButton from "./atoms/PrimaryButton"
 import { type PageInfo } from "../graphql/generated"
 import usePage from "../hooks/usePage"
 import { NextRouter } from "next/router"
