+++ pages/index.tsx
@@ -3,7 +3,7 @@ import Footer from "../components/Footer"
 import GameCard from "../components/GameCard"
 import NavBar from "../components/NavBar"
 import { useGamesQuery } from "../graphql/generated"
-import LoadingSpinner from "../components/LoadingSpinner"
+import LoadingSpinner from "../components/atoms/LoadingSpinner"
 import usePage from "../hooks/usePage"
 import PageNavigator from "../components/PageNavigator"
 import { useRouter } from "next/router"
